central_core_system/requirements.txt
# Central Core System Dependencies
# Decentralized AI Ecosystem - Central Server Requirements

# Web Framework and API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Message Queue and Task Processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Vector Database
chromadb==0.4.22

# LLM Integration
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.12
langgraph>=0.0.34

# Web Scraping and Search
requests==2.31.0
beautifulsoup4==4.12.2

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development and Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
